xwayland: Implement basic pointer warp emulation
When the client tries to warp the client, try to emulate the warping by
faking future pointer events until the pointer leaves the window. This
works by trying to lock the pointer on the surface, get relative
pointer, and use the relative pointer motions to move a fake cursor
position which is provided to the rest of the X server.

Signed-off-by: Jonas Ådahl <jadahl@gmail.com>