xwayland: Also provide unaccelerated when emulating pointer warping
In order for make it possible XI2 to provide raw unaccelerated pointer
motion data, pass both the unaccelerated and accelerated motion deltas.

Currently it is only possible to provide unaccelerated motion deltas
with relative pointer motions, so we cannot provide unaccelerated
or unobstructed relative motion deltas when we are not emulating
warping, since we always pass absolute pointer positions.

Signed-off-by: Jonas Ådahl <jadahl@gmail.com>