xwayland: Disable pointer warp emulation when cursor is shown
It is common that when a client shows the cursor after having warped
the pointer, it will no longer (at least for a period of time) continue
warping the pointer, and will expect it to work as normal. In order to
do what is most likely expected, disable the pointer lock when the
cursor is shown.

An example of this is a first-person-shooter game which shows the
cursor in order to let the player navigate a regular GUI with menus and
buttons etc.

Signed-off-by: Jonas Ådahl <jadahl@gmail.com>